requirements/base.txt
Django==2.0.6
Pillow==5.1.0
pytz==2018.4
raven==6.9.0
requests==2.19.1
django-allauth==0.36.0
django-cors-headers==2.2.1
django-extensions==2.0.7
django-filter==1.1.0
django-model-utils==3.1.2
django-nested-admin==3.0.21
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-simple-history==2.1.1
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
psycopg2==2.7.5
pycpfcnpj==1.4
python-dateutil==2.7.3
python-decouple==3.1
python-monkey-business==1.0.0
python3-openid==3.1.0
requests-oauthlib==1.0.0
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1


pdfminer.six==20181108
pycryptodome==3.7.3
sortedcontainers==2.1.0